Allowing random blur to be decimal
`PIL.ImageFilter.GaussianBlur` method allows decimal blur values, but using `random.randint` to generate a blur between 0 and the limit, will make the use of decimals generate an error, since you can't use an decimal in the randint method, and even if you could, it won't return decimal values. Then, the integer blur was always either too few or too agressive for TR. With this changes, it will be possible to find a number between these two extremes.